+++ lib/main_db/core/models/dto/dto.dart
@@ -1,0 +1,11 @@
+export 'category_dto.dart';
+export 'custom_field_dto.dart';
+export 'document_dto.dart';
+export 'file_metadata_dto.dart';
+export 'icon_dto.dart';
+export 'item_link_dto.dart';
+export 'recovery_code_dto.dart';
+export 'store_dto.dart';
+export 'tag_dto.dart';
+export 'vault_history_dto.dart';
+export 'vault_item_dto.dart';